[init] : change id type to Long

+++ src/main/java/com/looptracker/looptracker/repository/ITourInstanceRepository.java
@@ -3,5 +3,5 @@ package com.looptracker.looptracker.repository;
 import com.looptracker.looptracker.entity.TourInstance;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ITourInstanceRepository extends JpaRepository<TourInstance, Integer> {
+public interface ITourInstanceRepository extends JpaRepository<TourInstance, Long> {
 }
